Replaces clr with clr3
