+++ app/grants/management/commands/analytics_clr_old.py
@@ -19,10 +19,9 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 """
 
 from django.core.management.base import BaseCommand
-from django.db import connection
 from django.utils import timezone
 
-from grants.clr3 import calculate_clr, fetch_grants, get_summed_contribs_query, get_totals_by_pair, normalise
+from grants.clr_old import calculate_clr_for_donation, fetch_data, populate_data_for_clr
 from grants.models import GrantCLR
 
 
@@ -34,38 +33,26 @@ def analytics_clr(from_date=None, clr_round=None, network='mainnet'):
     # one-time data call
     total_pot = float(clr_round.total_pot)
     v_threshold = float(clr_round.verified_threshold)
+    uv_threshold = float(clr_round.unverified_threshold)
 
     print(total_pot)
 
-    grants = fetch_grants(clr_round, network)
+    grants, contributions = fetch_data(clr_round, network)
 
-    # collect contributions for clr_round into temp table
-    initial_query = get_summed_contribs_query(grants, clr_round.start_date, clr_round.end_date, clr_round.contribution_multiplier, network)
+    grant_contributions_curr = populate_data_for_clr(grants, contributions,  clr_round)
 
-    # open cursor and execute the groupBy sum for the round
-    with connection.cursor() as cursor:
-        curr_agg = {}
-        trust_dict = {}
-        # execute to populate shared state for the round
-        cursor.execute(initial_query) # (we could potential do better here by sharing this temp table between rounds)
-        for _row in cursor.fetchall():
-            if not curr_agg.get(_row[0]):
-                curr_agg[_row[0]] = {}
-
-            trust_dict[_row[1]] = _row[3]
-            curr_agg[_row[0]][_row[1]] = _row[2]
-
-        ptots = get_totals_by_pair(curr_agg)
-        bigtot, totals = calculate_clr(curr_agg, ptots, trust_dict, v_threshold, total_pot)
-
-        # normalise against a deepcopy of the totals to avoid mutations
-        curr_grants_clr = normalise(bigtot, totals, total_pot)
-        
-        # calculate clr analytics output
-        for grant in grants:
-            num_contribs, contrib_amount, clr_amount = curr_grants_clr.get(grant.id, {'num_contribs': 0, 'contrib_amount': 0, 'clr_amount': None}).values()
-            # debug_output.append([grant.id, grant.title, num_contribs, contrib_amount, clr_amount])
-            debug_output.append([grant.id, grant.title, grant.positive_round_contributor_count, float(grant.amount_received_in_round), clr_amount])
+    # calculate clr analytics output
+    for grant in grants:
+        clr_amount, _, num_contribs, contrib_amount = calculate_clr_for_donation(
+            grant,
+            0,
+            grant_contributions_curr,
+            total_pot,
+            v_threshold,
+            uv_threshold
+        )
+        # debug_output.append([grant.id, grant.title, num_contribs, contrib_amount, clr_amount])
+        debug_output.append([grant.id, grant.title, grant.positive_round_contributor_count, float(grant.amount_received_in_round), clr_amount])
 
     return debug_output
