+++ scripts/crontab2
@@ -1,4 +1,4 @@
 PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/usr/games:/usr/local/games:/snap/bin
 
 ## CLR
-# 1,30 * * * * cd gitcoin/coin; bash scripts/run_clr.bash slim  2>&1
+1,30 * * * * cd gitcoin/coin; bash scripts/run_clr.bash slim  2>&1
